Portal event invitation links to the website's needs form
Every event lead takes the website's write path: a WEB- reference, mount
site, in_progress until the questions are sent. The daily digest and the
booking notice read a portal lead like any other, and the row still names
the project through the query string the site writes onto it. The portal
route, the PTL- prefix and the portal-only configurator branches go.
