+++ echo/frontend/src/components/participant/DembraneEventCta.tsx
@@ -2,25 +2,21 @@ import { Trans } from "@lingui/react/macro";
 import { Box, Button } from "@mantine/core";
 import { IconArrowRight } from "@tabler/icons-react";
 import posthog from "posthog-js";
-import { useCallback } from "react";
 import PaulineUnderstandArt from "@/assets/pauline-understand.webp";
-import {
-	PricingConfigurator,
-	submitPortalConfiguration,
-	usePricingConfigurator,
-} from "@/components/pricing";
 import { useLanguage } from "@/hooks/useLanguage";
+import { eventEnquiryUrl } from "@/lib/links";
 import { testId } from "@/lib/testUtils";
 
 /** The last card on the thank you page: run your own event with dembrane.
  *
  * A participant has just spent ten minutes talking into their phone at
  * somebody else's event. This is the one moment they are a lead, so the card
- * uses one illustration and one button. The button opens the same
- * intake form the pricing configurator and the website's needs form use: an
- * email, six questions, then the event intake calendar, inside the portal
- * rather than on another site. The answers land on a `pricing_configuration`
- * row with a PTL- reference and the project the participant was in.
+ * uses one illustration and one button. The button opens the website's needs
+ * form in a new tab, with the project the participant was in on the URL. The
+ * site writes that onto the row as `project_id`, so the enquiry is a website
+ * enquiry like any other (a WEB- reference, `mount: site`, the daily digest
+ * and the booking notice) and still reads as "was at this event". One write
+ * path for every lead; nothing downstream has to know about the portal.
  *
  * Hosts on a paid plan can switch the card off per project
  * (`is_dembrane_event_cta_enabled`); the free tier always shows it. That
@@ -28,25 +24,7 @@ import { testId } from "@/lib/testUtils";
  */
 export const DembraneEventCta = ({ projectId }: { projectId: string }) => {
 	const { language } = useLanguage();
-	const configurator = usePricingConfigurator();
-
-	// The configurator's events, on the portal's own PostHog. `surface` tells
-	// the funnels apart from the dashboard's gates without a workspace.
-	const onEvent = useCallback(
-		(name: string, props: Record<string, unknown>) => {
-			posthog.capture(name, {
-				...props,
-				project_id: projectId,
-				surface: "participant_portal",
-			});
-		},
-		[projectId],
-	);
-
-	const handleOpen = () => {
-		posthog.capture("portal_event_cta_clicked", { project_id: projectId });
-		configurator.open();
-	};
+	const href = eventEnquiryUrl({ language, projectId });
 
 	return (
 		<>
@@ -65,27 +43,26 @@ export const DembraneEventCta = ({ projectId }: { projectId: string }) => {
 				{...testId("portal-finish-event-cta")}
 			>
 				<Button
+					component="a"
+					href={href}
+					rel="noopener noreferrer"
+					target="_blank"
 					// Tertiary on purpose. The page's job is done; this is a quiet
 					// offer, not a call to action, and it is already always in view.
 					variant="subtle"
 					size="lg"
 					fullWidth
 					rightSection={<IconArrowRight size={18} />}
-					onClick={handleOpen}
+					onClick={() =>
+						posthog.capture("portal_event_cta_clicked", {
+							project_id: projectId,
+						})
+					}
 					{...testId("portal-finish-event-cta-button")}
 				>
 					<Trans>dembrane at your event?</Trans>
 				</Button>
 			</Box>
-			<PricingConfigurator
-				{...configurator.configuratorProps}
-				entry="modal_direct"
-				locale={language}
-				mount="portal"
-				onEvent={onEvent}
-				projectId={projectId}
-				submit={submitPortalConfiguration}
-			/>
 		</>
 	);
 };
